Port ht from ADBench.
ADBench is a bit inconsistent about whether this benchmark is
abbreviated "HT" (for Hand Tracking) or "HAND". I have chosen the
former, as it is more consistent with the other names of ADBench
benchmarks.

Currently two tools implement this eval: PyTorch (implementation also
from ADBench) and Futhark. They validate against each other.

The run.py file for ht is currently hardcoded to only consider two
datasets for each of the four "classes", because the larger datasets
are extremely time consuming to run with PyTorch.