Implement ADBench benchmarks for Tapenade.
This PR begins the work of including Tapenade in Gradbench. Currently
only GMM is operational, but it does validate.

This work includes moving a lot of code from ADBench to Gradbench (and
also a JSON parser for C++). Some of the scaffolding code may
eventually be useful for other tools that are unfortunate enough to
speak C++.

The actual Gradbench-specific code is contained in the run_gmm.cpp
program.

Sadly, except for trivial evals (such as "hello"), running Tapenade
itself at define-time seems unviable, as the output code must in all
cases be manually modified in order to compile, and have various
supporting structures defined in seemingly ad-hoc ways. I have simply,
as with ADBench, included the thus-modified code.