Better error messages on validation errors.
This adds a new validator, gradbench.evaluation.mismatch, and a
function, gradbench.comparison.compare_json_objects, that produces
fairly human-friendly descriptions of mismatches between expected and
actual results.

All ADBench-derived evals have been modified to use it.

I had to modify the pytorch implementation of lstm, which had a small
bug where it returned a singleton list instead of a scalar (this was
not detected by the previous validation).