Add --runs to all ADBench-derived evals.
Also extends the PyTorch and Futhark tools to respect runs.

I had to perform some surgery on how timings are performed. Now,
rather than doing it in the generic run.py module itself, it is done
in each of the tool-specific eval modules. I think that is ultimately
the best place to do it, since it depends on the specific eval what is
to be measured, and how. The run function now returns both the result
and a list of timings (the latter which is just put into the response
message).