Construct the ADBench wrapper objects in prepare_input method.
Importantly, this moves the invocation of the "prepare" method out of
the timing loop. Depending on the PyTorch backend, "prepare" might be
very costly (e.g. copying data to a GPU), but is unrelated to AD.

A future dadition could be to also return timing information for
"prepare", in case anyone would find that interesting.