Support timeouts in intermediary.
This does work, although it has the following deficiencies:

1. Timeouts are only supported for tool responses, not eval responses.
This is probably not a significant problem.

2. Failures due to timeouts are not distinguishable from other
failures. It can be expected that some tools will fail to complete on
some datasets, and it would be nice if this did not cause CI to fail
entirely, but simply show up on the graphs (once we get those). I
think the solution here is to make the exit code depend on the reason
for the termination. Alternatively, just don't treat timeouts as an
error at all.

3. When run in the terminal, timeouts are somewhat noisy, because
pipes get broken and the tool/eval scripts get upset about unexpected
EOFs and SIGPIPEs and print out stack traces. Quite similar to what
happens when a tool crashes, actually.