Modify evals and tools to handle closed pipes gracefully.
Previously, if the stdin/stdout pipes were closed unexpectedly by the
intermediary (usually because of an error somewhere), we'd get loud
stack traces and nonzero exit codes. This has two problems:

* It obscures the first error (the one that causes the intermediary to
  close the pipes) because the followup error is just as verbose.

* It makes timeouts (#178) much more verbose, because they inherently
  depend on unexpectedly shutting down the tool and/or eval.

I think it is correct design for debuggability that the stderrs of
tool/eval are not in any way impeded, but it comes with an obligation
not to needlessly litter them. It is very unlikely that the change in
this PR will suppress a true bug.