Deduplicate Python modules for C++ tools.
The convention in Gradbench is that each eval for a tool is
represented by a Python module. This unfortunately means that the
C++-based tools have dozens of identical modules. This PR addresses
this by creating a single module that is then symlinked.

I also made the tools all run the compiler upon receiving 'define' -
previously some of them would do so in the Dockerfile itself. I am not
convinced this is actually the best approach (due to #170), but I
remember Sam preferring it this way.