Detect define and evaluation errors.
This draft PR extends the protocol definition and evaluation responses
with an "error" field that a tool can use to provide information on
what went wrong. This provides two important features:

* It allows us to distinguish a tool missing an eval from compilation
  of an eval failing.

* It allows evaluation failure to be reported as something beyond an
  invalid result.

Tools must be modified to actually catch exceptions (or whatever) and
stuff them into the responses). This is why this PR is marked as
draft: until #204 is merged, this means modifying a lot of
otherwise-identical modules.