Add "eval", "tool", and "config" fields to start message/response.
This PR adds a mandatory "eval" field to the "start" message, which
must contain the name of the eval.

It also adds a mandatory "tool" field to the tool response, which must
contain the name of the tool.

Finally, both the message and response may optionally contain a
"config" field that can contain arbitrary eval/tool-specific
configuration information.

This information is highly useful when writing scripts that process
log files. It means you need no information beyond what is in the file
itself. It is also very easy to implement, so imposes no real burden
on tool authors.