Replace --runs with --min-runs and --min-seconds.
The problem with --runs is that a fixed number of runs for all
workloads is not appropriate. You want many runs for small workloads,
and few for the big ones. This PR removes the --runs option from the
evals and replace it with two new options, with the following semantics:

  * --min-runs=N: run each workload at least N many times. Defaults to
    1, and must be greater than zero.

  * --min-seconds=N: run each workload until the aggregrate measured
    runtime is at least this many seconds. May be a fractional number.
    Defaults to 1.

This should not significantly increase time taken in CI (because it
will not by default rerun any workload that already takes a second),
but it should dramatically improve the reliability of runtime
measurements for smaller workloads.