New evals: particle, saddle
These are benchmarks from the 2016 paper *Efficient Implementation of
a Higher-Order Language with Built-In AD*:
https://www.bcl.hamilton.ie/~qobi/ad2016-benchmarks/

For some reason I am unable to find the paper that describes what
precisely they do, but from what I remember, 'particle' is about
simulating the movement of a particle in an electric field, and
'saddle' is about finding a saddle point for some function. Both make
use of the Euler method for doing so.

The compelling thing about these problems is that they involve nested
AD, and not just in the simple way needed by kmeans for computing
hessians. No, these things optimise objective functions that
themselves contain instances of AD (in the case of 'saddle', this is a
nested solver). The 'particle' eval even has unbounded 'while' loops
in the differentiated part. These ones are really looking for a fight.
Both have two levels of AD.

This PR is marked as a draft for various reasons, but the main one is
that these benchmarks don't fit real well in GradBench's current
model, so we need to figure out how to map them to the
module/function/workload hierarchy. For example, there is no
meaningful notion of a "primal" function here for which we can measure
AD overhead, like we do for all of the extant evals. But it gets
worse.

Each of the evals have two levels of AD, and each level can be
implemented with forward and reverse mode. Siskind and Pearlmutter
report performance numbers for each of the four combinations here:
https://engineering.purdue.edu/~qobi/stalingrad-examples2009/

I have chosen to model these four variants as four "functions", within
the same module.

In this PR, I only have a single tool implementation (Futhark), and I
implement each of the four variants for each of the two evals. This
works fine in Futhark, but I suspect some tools will have trouble
expressing some of these cases - doubly nested reverse mode with
'while' loops is *not* a trivial exercise. How will we handle the case
where a tool can only express a subset of these four combinations?

The functions have only a single workload each, taken from Siskind and
Pearlmutter's code distribution. In general, it seems scalability
across workloads was not a priority for Siskind and Pearlmutter, and I
arguably agree - there's not really any meaningful "input size" you
can dial here. One unfortunate consequence is that these problems are
vey sequential, but I think that's only a problem due to my own
research interests.

As alluded to above, these are complex uses of AD, and we should not
expect to easily obtain many tool implementations for these evals. I
am pretty sure Enzyme can do this (and I will give it a shot), and
maybe some of the other C++-tape-based ones (though I already dread
the template error messages). I was a bit surprised that Futhark
managed to handle it well. I don't think we can use any of the Siskind
and Pearlmutter code, as most of it involves implementing AD
facilities from scratch in the various languages, which is not
appropriate for GradBench. (Also, most of these languages are just
different Scheme implementations).