Port cmpad benchmarks
This currently only has `an_ode`, and only implemented for three
tools, but I intend to port all three benchmarks from cmpad, and as
many tools as I can.

Before anyone asks, I don't know why it is called `an_ode` and not
just `ode`, but I felt it was better to keep the original names.

There is a fair amount of additional mathematical information about
the benchmarks in the cmpad documentation. I have opted not to copy
it, for various reasons:

  * I am not confident in my ability to transcribe it without
    introducing errors.

  * GitHub's Markdown-math support is not particularly good.

  * The cmpad documentation is still necessary.

And honestly, the cmpad problems are not very exotic, so implementers
can find suitable information in many places.