julia: rewrite ode without side effects, use zygote
This PR rewrites the ode benchmark in a more array-oriented style
(based on the Futhark implementation), such that Zygote can handle it.
It is about x3 slower than the old imperative version, although I
suppose that could eventually be recovered by parallelising the
execution.

But there are questions here, which I think is best answered by the
honoured representatives of the Julia community:

1) Is this a more idiomatic implementation than the previous
imperative one?

2) How should we handle cases where one formulation of a benchmark is
advantageous for one tool, but disadvantageous for others? (Or simply
reduces primal performance.)

One possibility is to provide multiple implementations, and a tool can
then pick the one that works the best. This is already done for the
'det' benchmark, which is a rather extreme case of this situation.