cli: add 'log trim' command
This PR adds a new command to the gradbench CLI that sets the 'input'
and 'output' fields of a log file to null. The purpose of this is to
trim the often very large log files produced by GradBench, in the
interest of saving on storage space and reading time. For example, the
log file for gmm-futhark is reduced in size by about 95% after
trimming. Most of the remaining space is taken up by timings.

The use of null allows us to distinguish trimmed log files, as
presumably this value will never originally occur.

Although the command implementation is pretty short, I have factored
it into a separate file 'log.rs', because I envision that we will
eventually have a bunch of such log manipulation commands.

I am not a particularly experienced Rust programmer, so I would like
feedback on my approach before I start adding more subcommands.