Add CITATION.cff
This add a file containing metadata for citing GradBench. It is relevant because
I have some students who will need to cite GradBench properly soon-ish. The
citation file only covers GradBench as a software artifact, not any papers that
we will likely write based on it.

I simply added everyone who has commits in GradBench as an author, ordered by
their first commit. This seems the most fair. It cannot be fully automated,
because people will eventually make commits using different names, email
addresses, or anonymously. However, I don't expect maintenance of this file to
be particularly onerous. It is also not a permanent record, so we can add ORCIDs
and such over time.

If this is accepted, a good next step is to also upload GradBench to
[Zenodo](https://zenodo.org/). There are GitHub actions for automating this.